added inline

+++ src/threv-client.c
@@ -122,7 +122,7 @@ int ezudpcb (socket_t s, struct sockaddr_in *si_other, void *unused) {
 }
 
 int main (void) {
-   const int err = ezudp_client (1234, "127.0.0.1" /*"localhost"*/, ezudpcb, NULL);
+   const int err = ezudp_client ((uint16_t) 1234, "127.0.0.1" /*"localhost"*/, ezudpcb, NULL);
    if (err != 0) {
       fprintf (stderr, "err:%d\n", err);
       return EXIT_FAILURE;
